apiのリクエストファイル分け #41
