+++ frontend/react-app/src/lib/api/auth.ts
@@ -35,63 +35,3 @@ export const getCurrentUser = () => {
     }
   })
 }
-
-export const patientShow = () => {
-  return client.get("/patient_profiles/show", {
-    headers: {
-      "access-token": Cookies.get("_access_token")!,
-      "client": Cookies.get("_client")!,
-      "uid": Cookies.get("_uid")!
-    }
-  })
-}
-
-export const usersIndex = () => {
-  return client.get("/users/index", {
-    headers: {
-      "access-token": Cookies.get("_access_token")!,
-      "client": Cookies.get("_client")!,
-      "uid": Cookies.get("_uid")!
-    }
-  })
-}
-
-export const userShow = (id: number) => {
-  return client.get(`/users/show/?id=${id}`, {
-    headers: {
-      "access-token": Cookies.get("_access_token")!,
-      "client": Cookies.get("_client")!,
-      "uid": Cookies.get("_uid")!
-    }
-  })
-}
-
-export const interviewsIndex = () => {
-  return client.get("/interviews/index", {
-    headers: {
-      "access-token": Cookies.get("_access_token")!,
-      "client": Cookies.get("_client")!,
-      "uid": Cookies.get("_uid")!
-    }
-  })
-}
-
-export const interviewsIndexUser = (id: number) => {
-  return client.get(`/interviews/index?id=${id}`, {
-    headers: {
-      "access-token": Cookies.get("_access_token")!,
-      "client": Cookies.get("_client")!,
-      "uid": Cookies.get("_uid")!
-    }
-  })
-}
-
-export const interviewShow = (id: number) => {
-  return client.get(`/interviews/show?id=${id}`, {
-    headers: {
-      "access-token": Cookies.get("_access_token")!,
-      "client": Cookies.get("_client")!,
-      "uid": Cookies.get("_uid")!
-    }
-  })
-}
