+++ frontend/react-app/src/lib/api/interview.ts
@@ -1,0 +1,32 @@
+import client from "lib/api/client"
+import Cookies from "js-cookie"
+
+export const interviewsIndex = () => {
+  return client.get("/interviews/index", {
+    headers: {
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
+    }
+  })
+}
+
+export const interviewsIndexUser = (id: number) => {
+  return client.get(`/interviews/index?id=${id}`, {
+    headers: {
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
+    }
+  })
+}
+
+export const interviewShow = (id: number) => {
+  return client.get(`/interviews/show?id=${id}`, {
+    headers: {
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
+    }
+  })
+}
