+++ frontend/react-app/src/lib/api/patient.ts
@@ -1,0 +1,32 @@
+import client from "lib/api/client"
+import Cookies from "js-cookie"
+
+export const patientShow = () => {
+  return client.get("/patient_profiles/show", {
+    headers: {
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
+    }
+  })
+}
+
+export const usersIndex = () => {
+  return client.get("/users/index", {
+    headers: {
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
+    }
+  })
+}
+
+export const userShow = (id: number) => {
+  return client.get(`/users/show/?id=${id}`, {
+    headers: {
+      "access-token": Cookies.get("_access_token")!,
+      "client": Cookies.get("_client")!,
+      "uid": Cookies.get("_uid")!
+    }
+  })
+}
